% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{CCLENormalized}
\alias{CCLENormalized}
\title{Quantitative Proteomics of the Cancer Cell Line Encyclopedia}
\format{The data is an instance of class \code{MSnSet} from package \code{MSnbase}.}
\source{
\href{https://gygi.med.harvard.edu/publications/ccle}{Quantitative Proteomics of the Cancer Cell Line Encyclopedia}
}
\usage{
data(CCLENormalized)
}
\description{
In this dataset, we include quantitative proteome profiling by mass
spectrometry of 375 cell lines from Cancer Cell Line Encyclopedia. The
experiment was performed in multiplex format with 9 biological samples per
plex and one common sample to normalize between plexes. The normalized data
contains 42 multiplex experiments (TMT) which consists of 504 runs on the
mass spectrometer and over 1,500 hours of instrument time.

Ref: https://gygi.med.harvard.edu/publications/ccle
}
\details{

}
\examples{
library(msdata2)
data(CCLENormalized)
CCLENormalized
pData(CCLENormalized)
head(exprs(CCLENormalized))
}
\references{
\emph{Quantitative Proteomics of the Cancer Cell Line Encyclopedia.}
Nusinow, D., Szpyt, J., Ghandi, M., Rose, C., McDonald, E., Kalocsay, M.,
Jané-Valbuena, J., Gelfand, E., Schweppe, D., Jedrychowski, M., Golji, J.,
Porter, D., Rejtar, T., Wang, Y., Kryukov, G., Stegmeier, F., Erickson, B.,
Garraway, L., Sellers, W. and Gygi, S.
Cell. Volume 180, Issue 2, 23 January 2020, Pages 387-402.e16
doi: https://doi.org/10.1016/j.cell.2019.12.023.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/31978347}{PubMed})
}
\keyword{datasets}
