\name{pRolocdata}
\alias{pRolocdata}

\title{
List of pRolocdata data sets
}

\description{
  This function lists the data sets available in \code{msdata2}
  package by calling \code{data(package = "msdata2")}.
}

\usage{
msdata2()
}

\references{
  See in the respective data sets' manual pages for references to
  publications.
}
\author{
  Chong Tang <chong.tang.mat@gmail.com>
}
\examples{
msdata2()
}
