% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{msdata2metadata}
\alias{msdata2metadata}
\alias{print.msdata2metadata}
\title{Extract msdata2 metadata}
\usage{
msdata2metadata(x)
}
\arguments{
\item{x}{A \code{msdata2} data.}
}
\value{
An instance of class \code{msdata2metadata}.
}
\description{
Extracts relevant metadata from an \code{MSnSet} instance. See
\code{README.md} for a description and explanation of the metadata
fields.
}
\examples{
library("msdata2")
data(Rost2014sgsHuman)
data(Rost2014sgsHuman)
msdata2metadata(Rost2014sgsHuman)
}
\author{
Chong Tang
}
