% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{ramus2016}
\alias{ramus2016}
\alias{ramus2016SCMfPaQ}
\alias{ramus2016SCMaxQuant}
\alias{ramus2016SCIrmaHeidi}
\alias{ramus2016SCScaffold}
\alias{ramus2016IntMFPaQ}
\alias{ramus2016IntMaxQuant}
\alias{ramus2016LFQMaxQuant}
\alias{ramus2016IntSkyline}
\title{Spiked Proteomic Standard Datasets For Benchmark 8 Different
label-free Quantitative Workflows}
\format{The data is an instance of class \code{MSnSet} from package \code{MSnbase}.}
\source{
This data was generated from Sup Table 1. Quantitative data obtained from
the 8 different workflows. See inst/scripts/ramus2016.R for details.
}
\usage{
data(ramus2016SCMfPaQ)
data(ramus2016SCMaxQuant)
data(ramus2016SCIrmaHeidi)
data(ramus2016SCScaffold)
data(ramus2016IntMFPaQ)
data(ramus2016IntMaxQuant)
data(ramus2016LFQMaxQuant)
data(ramus2016IntSkyline)
}
\description{
This data consists of quantitative results from 8 different Label-Free
quantitative workflows for a controlled, spiked proteomic standard dataset.
The proteomics standard data was prepared using a yeast cell lysate
accompanied by a serial dilution (respectively 0.05-0.125-0.250-0.5-2.5-5-
12.5-25-50 fmol of UPS1 \eqn{/\mu}g of yeast lysate) of the UPS1 standard mixture
(Sigma). Samples were then analyzed in triplicate by nanoLC-MS/MS coupled to
an LTQ-Orbitrap Velos mass spectrometer, operated in data-dependent
acquisition mode.

The dataset was then processed by 8 different workflows for benchmarking,
consisting in the following steps: peaklist generation, database search,
validation of the identified proteins and extraction of quantitative metric
(spectral count or MS signal).
}
\details{
Full details of the experimental design can be found in the reference.\cr

The 8 different datasets correspond to 8 different LFQ workflows:\cr
}
\section{ramus2016SCMfPaQ}{

             \verb{}Peaklist creation device - ExtractMSn\cr
                    Database search engine - Mascot\cr
                    Validation/Spectral counting device - MFPaQ\cr
                    Quantification method - Spectral counting\cr
}

\section{ramus2016SCMaxQuant}{

                    Peaklist creation device - Andromeda\cr
                    Database search engine - Andromeda\cr
                    Validation/Spectral counting device - MaxQuant\cr
                    Quantification method - Spectral counting\cr
}

\section{ramus2016SCIrmaHeidi}{

                    Peaklist creation device - Mascot Distiller\cr
                    Database search engine - Mascot\cr
                    Validation/Spectral counting device - IRMa/hEIDI\cr
                    Quantification method - Spectral counting\cr
}

\section{ramus2016SCScaffold}{

                    Peaklist creation device - ExtractMSn\cr
                    Database search engine - Mascot\cr
                    Validation/Spectral counting device - Scaffold\cr
                    Quantification method - Spectral counting\cr
}

\section{ramus2016IntMFPaQ}{

                    Peaklist creation device - ExtractMSn\cr
                    Database search engine - Mascot\cr
                    Validation/Spectral counting device - MFPaQ\cr
                    MS signal extraction device - MFPaQ\cr
                    Quantification method - MS signal analysis\cr
}

\section{ramus2016IntMaxQuant}{

                    Peaklist creation device - Andromeda\cr
                    Database search engine - Andromeda\cr
                    Validation/Spectral counting device - MaxQuant\cr
                    MS signal extraction device - MaxQuant (Intensity)\cr
                    Quantification method - MS signal analysis\cr
}

\section{ramus2016LFQMaxQuant}{

                    Peaklist creation device - Andromeda\cr
                    Database search engine - Andromeda\cr
                    Validation/Spectral counting device - MaxQuant\cr
                    MS signal extraction device - MaxQuant (LFQ)\cr
                    Quantification method - MS signal analysis\cr
}

\section{ramus2016IntSkyline}{

                    Peaklist creation device - Mascot Distiller\cr
                    Database search engine - Mascot\cr
                    Validation/Spectral counting device - Scaffold\cr
                    MS signal extraction device - Skyline\cr
                    Quantification method - MS signal analysis\cr
}

\examples{
library(msdata2)
data(ramus2016IntSkyline)
ramus2016IntSkyline
pData(ramus2016IntSkyline)
head(exprs(ramus2016IntSkyline))
}
\references{
\emph{Benchmarking quantitative label-free LC-MS data
processing workflows using a complex spiked proteomic standard dataset.}
Ramus C, Hovasse A, Marcellin M, Hesse AM, Mouton-Barbosa E, Bouyssié D,
Vaca S, Carapito C, Chaoui K, Bruley C, Garin J, Cianférani S, Ferro M,
Van Dorssaeler A, Burlet-Schiltz O, Schaeffer C, Couté Y, Gonzalez de Peredo A.
J Proteomics. 2016 Jan 30;132:51-62. \cr
doi: 10.1016/j.jprot.2015.11.011.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26585461}{PubMed})
}
\keyword{datasets}
