% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{msdata2}
\alias{msdata2}
\title{msdata2}
\usage{
msdata2()
}
\description{
This function lists the data sets available in \code{msdata2}
  package by calling \code{data(package = "msdata2")}
}
\references{
See in the respective data sets' manual pages for references to
publications.
}
\author{
Chong Tang <chong.tang.mat@gmail.com>
}
